backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-cloud-aiplatform
pydub
numpy
python-dotenv
google-genai
